Make redux actions type safe
