+++ src/core/redux/hooks.ts
@@ -5,17 +5,17 @@ import {
   useStore as useReduxStore,
 } from 'react-redux';
 
-import { Dispatch, State, Thunk } from './Store';
+import { Dispatch, State, Store, Thunk } from './types';
 
 /**
  * Expand of the base hooks of Redux by adding types.
  */
 
-export const useDispatch = (...args: Parameters<typeof useReduxDispatch>) => useReduxDispatch<Dispatch>(...args);
+export const useDispatch: () => Dispatch = useReduxDispatch;
 
 export const useSelector: TypedUseSelectorHook<State> = useReduxSelector;
 
-export const useStore = (...args: Parameters<typeof useReduxStore>) => useReduxStore(...args);
+export const useStore: () => Store = useReduxStore;
 
 export const useThunk = <TReturn = void>(thunk: ReturnType<Thunk<TReturn>>) => {
   const dispatch = useDispatch();
