+++ src/core/redux/reducer.ts
@@ -1,0 +1,11 @@
+import { combineReducers } from 'redux';
+
+import { PaymentAction, paymentsReducer } from 'payments/reducers';
+import { ShopAction, shopReducer } from 'shop/reducers';
+
+export type Action = PaymentAction | ShopAction;
+
+export const rootReducer = combineReducers({
+  payments: paymentsReducer,
+  shop: shopReducer,
+});
